requirements.txt
langgraph>=0.1.0
langchain>=0.1.0
langchain-community>=0.1.0
langchain-ollama>=0.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
